requirements.txt
asn1crypto==0.24.0
astroid==2.3.1
attrs==19.1.0
Automat==0.7.0
beautifulsoup4==4.8.1
bs4==0.0.1
cachetools==4.0.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
constantly==15.1.0
cryptography==2.7
cssselect==1.1.0
google-api-python-client==1.7.11
google-auth==1.10.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.15.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.21
lazy-object-proxy==1.4.2
lxml==4.4.1
mccabe==0.6.1
oauthlib==3.1.0
parsel==1.5.2
prettytable==0.7.2
psycopg2==2.8.3
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pylint==2.4.2
pyOpenSSL==19.0.0
pytube==9.5.3
queuelib==1.5.0
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
Scrapy==1.7.3
service-identity==18.1.0
six==1.12.0
soupsieve==1.9.4
Twisted==19.7.0
typed-ast==1.4.0
uritemplate==3.0.1
urllib3==1.25.6
w3lib==1.21.0
wrapt==1.11.2
youtube-mp3-downloader==1.4
zope.interface==4.6.0
moviepy
python-twitter
